Split ProgressInternal.cs to multiple files.
