+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/Progress.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05bad5b7396c0b64cba5a1678d664d42
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
