+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/Progress/ProgressMergeInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbe6538ff6100fc45b0a47866c48ef12
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
