+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/Progress/ProgressMultiAwaitInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f5ad516bfb142e4587a33a181980a9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
