+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/Progress/ProgressRaceInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44decb14347d58c47beb80635db09b51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
