+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/Progress/ProgressSingleAwaitInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca8db309bdef789499231c8cb1c55966
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
